Port over changes from XFree86/Mesa 5.0.2, mostly to silence compiler warnings.

+++ src/mesa/main/extensions.c
@@ -31,7 +31,7 @@
 #include "mtypes.h"
 
 
-#define F(x) (int)&(((struct gl_extensions *)0)->x)
+#define F(x) (int)(unsigned long)&(((struct gl_extensions *)0)->x)
 #define ON GL_TRUE
 #define OFF GL_FALSE
